+++ install/migrations/update_10.0.16_to_10.0.17/rule_tickets.php
@@ -33,30 +33,19 @@
  * ---------------------------------------------------------------------
  */
 
-namespace tests\units\Glpi\ContentTemplates\Parameters;
+/**
+ * @var \DBmysql $DB
+ * @var \Migration $migration
+ */
 
-class KnowbaseItemParameters extends AbstractParameters
-{
-    public function testGetValues(): void
-    {
-        $this->login();
-
-        $this->createItem('KnowbaseItem', [
-            'name'        => 'kbi_testGetValues',
-            'answer'      => "test answer' \"testGetValues",
-        ]);
-
-        $kbi_id = getItemByTypeName('KnowbaseItem', 'kbi_testGetValues', true);
-
-        $parameters = $this->newTestedInstance();
-        $values = $parameters->getValues(getItemByTypeName('KnowbaseItem', 'kbi_testGetValues'));
-        $this->array($values)->isEqualTo([
-            'id'     => $kbi_id,
-            'name'   => 'kbi_testGetValues',
-            'answer' => "test answer' \"testGetValues",
-            'link'   => '<a href="/glpi/front/knowbaseitem.form.php?id=' . $kbi_id . '" title="kbi_testGetValues">kbi_testGetValues</a>',
-        ]);
-
-        $this->testGetAvailableParameters($values, $parameters->getAvailableParameters());
-    }
-}
+$migration->addPostQuery(
+    $DB->buildUpdate(
+        'glpi_rulecriterias',
+        [
+            'criteria' => 'users_id_recipient',
+        ],
+        [
+            'criteria' => '_users_id_recipient',
+        ]
+    )
+);