+++ src/Rack.php
@@ -435,8 +435,8 @@ class Rack extends CommonDBTM
         );
 
         echo "<div id='switchview'>";
-        echo "<i id='sviewlist' class='pointer ti ti-list' title='" . __('View as list') . "'></i>";
-        echo "<i id='sviewgraph' class='pointer ti ti-layout-grid selected' title='" . __('View graphical representation') . "'></i>";
+        echo "<i id='sviewlist' class='pointer ti ti-list' title='" . __s('View as list') . "'></i>";
+        echo "<i id='sviewgraph' class='pointer ti ti-layout-grid selected' title='" . __s('View graphical representation') . "'></i>";
         echo "</div>";
 
         $racks = iterator_to_array($racks);
@@ -444,7 +444,7 @@ class Rack extends CommonDBTM
 
         $rack = new self();
         if (!count($racks)) {
-            echo "<table class='tab_cadre_fixe'><tr><th>" . __('No rack found') . "</th></tr>";
+            echo "<table class='tab_cadre_fixe'><tr><th>" . __s('No rack found') . "</th></tr>";
             echo "</table>";
         } else {
             if ($canedit) {
@@ -463,7 +463,7 @@ class Rack extends CommonDBTM
                 $header .= Html::getCheckAllAsCheckbox('mass' . __CLASS__ . $rand);
                 $header .= "</th>";
             }
-            $header .= "<th>" . __('Name') . "</th>";
+            $header .= "<th>" . __s('Name') . "</th>";
             $header .= "</tr>";
 
             echo $header;
@@ -532,7 +532,7 @@ class Rack extends CommonDBTM
 
         if (count($outbound)) {
             echo "<table class='outbound'><thead><th>";
-            echo __('Following elements are out of room bounds');
+            echo __s('Following elements are out of room bounds');
             echo "</th></thead><tbody>";
             foreach ($outbound as $out) {
                 $rack->getFromResultSet($out);
@@ -972,33 +972,33 @@ JAVASCRIPT;
      */
     private static function getCell(Rack $rack, $cell)
     {
-        $bgcolor = $rack->getField('bgcolor');
-        $fgcolor = Html::getInvertedColor($bgcolor);
-        return "<div class='grid-stack-item room_orientation_" . $cell['room_orientation'] . "'
-                  gs-id='" . $cell['id'] . "'
+        $bgcolor = htmlspecialchars($rack->getField('bgcolor'));
+        $fgcolor = htmlspecialchars(Html::getInvertedColor($bgcolor));
+        return "<div class='grid-stack-item room_orientation_" . htmlspecialchars($cell['room_orientation']) . "'
+                  gs-id='" . htmlspecialchars($cell['id']) . "'
                   gs-locked='true'
                   gs-h='1'
                   gs-w='1'
-                  gs-x='" . $cell['_x'] . "'
-                  gs-y='" . $cell['_y'] . "'>
+                  gs-x='" . htmlspecialchars($cell['_x']) . "'
+                  gs-y='" . htmlspecialchars($cell['_y']) . "'>
             <div class='grid-stack-item-content'
                   style='background-color: $bgcolor;
                         color: $fgcolor;'>
                <a href='" . $rack->getLinkURL() . "'
                   style='color: $fgcolor'>" .
-                  $cell['name'] . "</a>
+                  htmlspecialchars($cell['name']) . "</a>
                <span class='tipcontent'>
                   <span>
-                     <label>" . __('name') . ":</label>" .
-                     $cell['name'] . "
+                     <label>" . __s('name') . ":</label>" .
+                     htmlspecialchars($cell['name']) . "
                   </span>
                   <span>
-                     <label>" . __('serial') . ":</label>" .
-                     $cell['serial'] . "
+                     <label>" . __s('serial') . ":</label>" .
+                     htmlspecialchars($cell['serial']) . "
                   </span>
                   <span>
-                     <label>" . __('Inventory number') . ":</label>" .
-                     $cell['otherserial'] . "
+                     <label>" . __s('Inventory number') . ":</label>" .
+                     htmlspecialchars($cell['otherserial']) . "
                   </span>
                </span>
             </div><!-- // .grid-stack-item-content -->