+++ tests/functional/Glpi/ContentTemplates/Parameters/ProblemParameters.php
@@ -69,7 +69,7 @@ class ProblemParameters extends AbstractParameters
         $this->array($values)->isEqualTo([
             'id'        => $problems_id,
             'ref'       => "#$problems_id",
-            'link'      => "<a  href='/glpi/front/problem.form.php?id=$problems_id'  title=\"problem_testGetValues\">problem_testGetValues</a>",
+            'link'      => '<a href="/glpi/front/problem.form.php?id=' . $problems_id . '" title="problem_testGetValues">problem_testGetValues</a>',
             'name'      => 'problem_testGetValues',
             'content'   => '<p>problem_testGetValues content</p>',
             'date'      => '2021-07-19 17:11:28',