+++ tests/functional/Glpi/ContentTemplates/Parameters/TicketParameters.php
@@ -116,8 +116,9 @@ class TicketParameters extends AbstractParameters
         ]);
         $tickets_id = $created_ticket->getID();
 
+        $kb_item_id = getItemByTypeName(\KnowbaseItem::class, '_knowbaseitem01', true);
         $this->createItem(\KnowbaseItem_Item::class, [
-            'knowbaseitems_id' => getItemByTypeName(\KnowbaseItem::class, '_knowbaseitem01', true),
+            'knowbaseitems_id' => $kb_item_id,
             'itemtype'         => 'Ticket',
             'items_id'         => $tickets_id,
         ]);
@@ -127,7 +128,7 @@ class TicketParameters extends AbstractParameters
         $this->array($values)->isEqualTo([
             'id'        => $tickets_id,
             'ref'       => "#$tickets_id",
-            'link'      => "<a  href='/glpi/front/ticket.form.php?id=$tickets_id'  title=\"ticket_testGetValues\">ticket_testGetValues</a>",
+            'link'      => '<a href="/glpi/front/ticket.form.php?id=' . $tickets_id . '" title="ticket_testGetValues">ticket_testGetValues</a>',
             'name'      => 'ticket_testGetValues',
             'content'   => '<p>ticket_testGetValues content</p>',
             'date'      => $now,
@@ -254,10 +255,10 @@ class TicketParameters extends AbstractParameters
             ],
             'knowbaseitems' => [
                 [
-                    'id' => getItemByTypeName(\KnowbaseItem::class, '_knowbaseitem01', true),
+                    'id' => $kb_item_id,
                     'name' => '_knowbaseitem01',
                     'answer' => "Answer for Knowledge base entry _knowbaseitem01 apple juice turnover",
-                    'link' => "<a  href='/glpi/front/knowbaseitem.form.php?id=1'  title=\"_knowbaseitem01\">_knowbaseitem01</a>"
+                    'link' => '<a href="/glpi/front/knowbaseitem.form.php?id=' . $kb_item_id . '" title="_knowbaseitem01">_knowbaseitem01</a>'
                 ]
             ],
             'assets'        => [],